// model/testData.types.ts
export type LoginData = {
  username: string;
  password: string;
};

export type UserData = {
  role:string;
  status:string;
  employeeName: string;
  username: string;
  updatedUsername: string;
  password: string;
};

export type EmployeeData = {
  firstName: string;
  middleName: string;
  lastName: string;
  employeeId: string;
  updatedFirstName: string;
  updatedMiddleName: string;
  updatedLastName: string;
  otherID: string;
};